EXPORT_PLATFORM env added

+++ src/setup.sh
@@ -21,7 +21,7 @@ TRES_PATH=$ROOT_PATH/.config/godot/editor_settings-3.tres
 # GODOT_DL_SUBDIR=$GODOT_VERSION
 
 echo "✔ Validating Environment Variables."
-echo ${GODOT_VERSION} ${GODOT_RELEASE} ${ROOT_PATH}
+echo ${GODOT_VERSION} ${GODOT_RELEASE} ${ROOT_PATH} ${EXPORT_PLATFORM}
 printenv
 
 # Download and Install Packages
@@ -55,41 +55,45 @@ sudo rm -f Godot_v${GODOT_VERSION}-${GODOT_RELEASE}_export_templates.tpz Godot_v
 
 echo "✔ Engine & Export Templates Successfully Installed."
 
-# Android SDK
-sudo mkdir -p -v /root/android-sdk-installer/cmdline-tools && cd /root/android-sdk-installer/cmdline-tools
-curl -fsSLO "https://dl.google.com/android/repository/commandlinetools-linux-6858069_latest.zip"
-unzip -q commandlinetools-linux-*.zip
-rm commandlinetools-linux-*.zip
-mv cmdline-tools latest
-mkdir -p -v /root/.android && echo "count=0" > /root/.android/repositories.cfg
-yes | /root/android-sdk-installer/cmdline-tools/latest/bin/sdkmanager --licenses
-yes | /root/android-sdk-installer/cmdline-tools/latest/bin/sdkmanager --sdk_root=$ANDROID_HOME "platform-tools" "build-tools;30.0.3" "platforms;android-29" "cmdline-tools;latest" "cmake;3.10.2.4988404" "ndk;21.4.7075529"
-cd /root && rm -rf /root/android-sdk-installer
-
-echo "✔ Android SDK Successfully Installed."
-
-# Key Generation
-keytool -keyalg RSA -genkeypair -alias androiddebugkey -keypass android -keystore /root/android-sdk/debug.keystore -storepass android -dname "CN=Android Debug,O=Android,C=US" -validity 9999 
-ls /root/android-sdk/debug.keystore
-
-echo "✔ Debug Key Generated."
+# Android Export
+if [ "${EXPORT_PLATFORM}" == "android" ]
+then 
+    # Android SDK
+    sudo mkdir -p -v /root/android-sdk-installer/cmdline-tools && cd /root/android-sdk-installer/cmdline-tools
+    curl -fsSLO "https://dl.google.com/android/repository/commandlinetools-linux-6858069_latest.zip"
+    unzip -q commandlinetools-linux-*.zip
+    rm commandlinetools-linux-*.zip
+    mv cmdline-tools latest
+    mkdir -p -v /root/.android && echo "count=0" > /root/.android/repositories.cfg
+    yes | /root/android-sdk-installer/cmdline-tools/latest/bin/sdkmanager --licenses
+    yes | /root/android-sdk-installer/cmdline-tools/latest/bin/sdkmanager --sdk_root=$ANDROID_HOME "platform-tools" "build-tools;30.0.3" "platforms;android-29" "cmdline-tools;latest" "cmake;3.10.2.4988404" "ndk;21.4.7075529"
+    cd /root && rm -rf /root/android-sdk-installer    
+    echo "✔ Android SDK Successfully Installed."
+    # Key Generation
+    keytool -keyalg RSA -genkeypair -alias androiddebugkey -keypass android -keystore /root/android-sdk/debug.keystore -storepass android -dname "CN=Android Debug,O=Android,C=US" -validity 9999 
+    ls /root/android-sdk/debug.keystore
+    echo "✔ Debug Key Generated."
+fi
 
 # Godot Executable From Path
 #!/usr/bin/godot
 chmod +x /usr/local/bin/godot && godot -q
 
 echo "✔ Godot Editor First Launch."
-# Set New Editor Settings
-sed -i '/\[resource\]/a export\/android\/android_sdk_path = "/root/android-sdk"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/adb = "/root/android-sdk/platform-tools/adb"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/jarsigner = "/usr/bin/jarsigner"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/apksigner = "/usr/bin/apksigner"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/debug_keystore = "/root/android-sdk/debug.keystore"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/debug_user = "androiddebugkey"' ${TRES_PATH} \
-&& sed -i '/\[resource\]/a export\/android\/debug_pass = "android"' ${TRES_PATH}
+
+if [ "${EXPORT_PLATFORM}" == "android" ]
+then 
+    # Set New Editor Settings
+    sed -i '/\[resource\]/a export\/android\/android_sdk_path = "/root/android-sdk"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/adb = "/root/android-sdk/platform-tools/adb"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/jarsigner = "/usr/bin/jarsigner"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/apksigner = "/usr/bin/apksigner"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/debug_keystore = "/root/android-sdk/debug.keystore"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/debug_user = "androiddebugkey"' ${TRES_PATH} \
+    && sed -i '/\[resource\]/a export\/android\/debug_pass = "android"' ${TRES_PATH}
+    echo "✔ Android Project Export Setup Ready"
+fi
 
 cat ${TRES_PATH}
 # cd /game && mkdir -v -p build/android
-# godot --verbose --export-debug "Android" build/android/$EXPORT_NAME.debug.apk
-
-echo "✔ Android Project Export Setup Ready"
+# godot --verbose --export-debug ${EXPORT_PLATFORM} build/android/$EXPORT_NAME.debug.apk
